requirements.txt
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
executing==2.0.1
filelock==3.13.3
Flask==3.0.2
gunicorn==22.0.0
icecream==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pipenv==2023.12.1
platformdirs==4.2.0
Pygments==2.17.2
pymongo==4.6.2
python-dotenv==1.0.1
six==1.16.0
virtualenv==20.25.1
webencodings==0.5.1
Werkzeug==3.0.1
